fix: 음료 주문 페이지 개선
